+++ tests/snapshots/test_snapshot_worker.py
@@ -10,7 +10,7 @@ from snuba.datasets.factory import get_dataset
 from snuba.processor import ProcessorAction, ProcessedMessage
 from snuba.stateful_consumer.control_protocol import TransactionData
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.consumer import KafkaMessage, Partition, Topic
+from snuba.utils.streams.types import Message, Partition, Topic
 from tests.backends.confluent_kafka import FakeConfluentKafkaProducer
 
 
@@ -79,6 +79,6 @@ class TestSnapshotWorker:
         )
 
         ret = worker.process_message(
-            KafkaMessage(Partition(Topic("topic"), 0), 1, message.encode("utf-8"),)
+            Message(Partition(Topic("topic"), 0), 1, message.encode("utf-8"),)
         )
         assert ret == expected
