+++ tests/test_consumer.py
@@ -6,7 +6,7 @@ from snuba.consumer import ConsumerWorker
 from snuba.datasets.factory import enforce_table_writer
 from snuba.processor import ProcessedMessage, ProcessorAction
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.consumer import KafkaMessage, Partition, Topic
+from snuba.utils.streams.types import Message, Partition, Topic
 from tests.base import BaseEventsTest
 from tests.backends.confluent_kafka import FakeConfluentKafkaProducer
 
@@ -18,7 +18,7 @@ class TestConsumer(BaseEventsTest):
     def test_offsets(self):
         event = self.event
 
-        message = KafkaMessage(
+        message = Message(
             Partition(Topic("events"), 456),
             123,
             json.dumps((0, "insert", event)).encode(
@@ -64,7 +64,7 @@ class TestConsumer(BaseEventsTest):
         event["data"]["datetime"] = old_timestamp_str
         event["data"]["received"] = int(calendar.timegm(old_timestamp.timetuple()))
 
-        message = KafkaMessage(
+        message = Message(
             Partition(Topic("events"), 1),
             42,
             json.dumps((0, "insert", event)).encode("utf-8"),
