+++ tests/utils/streams/test_batching.py
@@ -12,13 +12,13 @@ from typing import (
 from unittest.mock import patch
 
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.consumer import KafkaMessage, Partition, Topic
 from snuba.utils.streams.batching import AbstractBatchWorker, BatchingConsumer
+from snuba.utils.streams.types import Message, Partition, Topic
 
 
 class FakeKafkaConsumer:
     def __init__(self):
-        self.items: MutableSequence[KafkaMessage] = []
+        self.items: MutableSequence[Message] = []
         self.commit_calls = 0
         self.close_calls = 0
         self.positions: MutableMapping[Partition, int] = {}
@@ -34,7 +34,7 @@ class FakeKafkaConsumer:
     def unsubscribe(self) -> None:
         pass  # XXX: This is a bit of a smell.
 
-    def poll(self, timeout: Optional[float] = None) -> Optional[KafkaMessage]:
+    def poll(self, timeout: Optional[float] = None) -> Optional[Message]:
         try:
             message = self.items.pop(0)
         except IndexError:
@@ -69,7 +69,7 @@ class FakeWorker(AbstractBatchWorker[Any]):
         self.processed: MutableSequence[Optional[Any]] = []
         self.flushed: MutableSequence[Sequence[Any]] = []
 
-    def process_message(self, message: KafkaMessage) -> Optional[Any]:
+    def process_message(self, message: Message) -> Optional[Any]:
         self.processed.append(message.value)
         return message.value
 
@@ -91,7 +91,7 @@ class TestConsumer(object):
         )
 
         consumer.items = [
-            KafkaMessage(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
+            Message(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
             for i in [1, 2, 3]
         ]
         for x in range(len(consumer.items)):
@@ -118,7 +118,7 @@ class TestConsumer(object):
 
         mock_time.return_value = time.mktime(datetime(2018, 1, 1, 0, 0, 0).timetuple())
         consumer.items = [
-            KafkaMessage(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
+            Message(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
             for i in [1, 2, 3]
         ]
         for x in range(len(consumer.items)):
@@ -126,7 +126,7 @@ class TestConsumer(object):
 
         mock_time.return_value = time.mktime(datetime(2018, 1, 1, 0, 0, 1).timetuple())
         consumer.items = [
-            KafkaMessage(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
+            Message(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
             for i in [4, 5, 6]
         ]
         for x in range(len(consumer.items)):
@@ -134,7 +134,7 @@ class TestConsumer(object):
 
         mock_time.return_value = time.mktime(datetime(2018, 1, 1, 0, 0, 5).timetuple())
         consumer.items = [
-            KafkaMessage(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
+            Message(Partition(Topic("topic"), 0), i, f"{i}".encode("utf-8"))
             for i in [7, 8, 9]
         ]
         for x in range(len(consumer.items)):
